feat: refactor character model and update imports across components for better organization
