+++ .gitignore
@@ -23,6 +23,7 @@ pids
 /node_modules
 npm-debug.log
 yarn-error.log
+package-lock.json
 
 # Dependency directories
 node_modules/
