+++ src/app/core/models/character.model.ts
@@ -1,0 +1,13 @@
+export interface Character {
+  id: number;
+  name: string;
+  status: string;
+  species: string;
+  gender: string;
+  type: string;
+  image: string;
+  location: {
+    name: string;
+    url: string;
+  };
+}
