+++ src/app/features/home/components/character-section/character-section.component.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
-import { Character, CharacterService } from '@core/services/character.service';
+import { Character } from '@core/models/character.model';
+import { CharacterService } from '@core/services/character.service';
 import { CardCharacterComponent } from '@shared/components/card-character/card-character.component';
 
 @Component({
